Issue #121: Safe entity delete in Drupal 8 or higher

+++ src/Behat/Cores/Drupal8.php
@@ -360,7 +360,9 @@ class Drupal8 extends OriginalDrupal8 implements CoreInterface {
     }
 
     $entity = $controller->load($entity_id);
-    $entity->delete();
+    if ($entity instanceof EntityInterface) {
+      $entity->delete();
+    }
   }
 
   /**
